Login/signup setup

+++ backend/controllers/sauceControllers.js
@@ -1,25 +1,56 @@
 const Sauce = require('../models/sauce');
+const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
-  delete req.body._id;
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
+  delete sauceObject._userId;
   const sauce = new Sauce({
-    ...req.body
+    ...sauceObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   });
+
   sauce.save()
-    .then(() => res.status(201).json({message: 'Sauce enregistrée'}))
-    .catch(error => res.status(400).json({error}));
+  .then(() => {res.status(201).json({message: 'Sauce eenregistrée !'})})
+  .catch(error => res.status(400).json({error}));
 };
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-      .then(() => res.status(200).json({message: 'Sauce modifiée !'}))
+    const sauceObject = req.file ? {
+      ...JSON.parse(req.body.sauce),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : {...req.body};
+
+    delete sauceObject._userId;
+    Sauce.findOne({_id: req.params.id})
+      .then(sauce => {
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({message: 'Not authorized'});
+        } else {
+          Sauce.updateOne({_id: req.params.id}, {...sauceObject, _id: req.params.id})
+          .then(() => res.status(200).json({message: 'Sauce modifiée !'}))
+          .catch(error => res.status(401).json({error}));
+        }
+      })
       .catch(error => res.status(400).json({error}));
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-      .then(() => res.status(200).json({message: 'Sauce supprimée !'}))
-      .catch(error => res.status(400).json({error}));
+    Sauce.findOne({_id: req.params.id})
+      .then(sauce => {
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({message: 'Not authorized'})
+        } else {
+          const filename = sauce.imageUrl.splut('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.deleteOne({_id: req.params.id})
+              .then(() => {res.status(200).json({message: 'Sauce supprimée )'})})
+              .catch(error => res.status(401).json({error}));
+          });
+        }
+      })
+      .catch(error => res.status(500).json({error}));
 };
 
 exports.getOneSauce = (req, res, next) => {
